fix: extensions input query updated
