+++ extensions/product-discount/src/shopify-discount-meta.ts
@@ -1,0 +1,12 @@
+export interface ShopifyDiscountMeta {
+    type: string;
+    description: string;
+    discountType: string;
+    discountValue: number;
+    minValue: number;
+    minQty: number;
+    maxUse: number;
+    onePerUser: boolean;
+    products: string[];
+    collections: string[];
+  }
